Split into two crates, since ["staticlib", "rlib"] causes both slowness and linking issues.
