+++ benchmarks/results/lots-of-peaks.json
@@ -1,12 +1,12 @@
 {
-    "D1mr": 4648862,
-    "D1mw": 2816797,
-    "DLmr": 1568319,
-    "DLmw": 1792029,
-    "Dr": 302525899,
-    "Dw": 149719569,
-    "I1mr": 7252806,
-    "ILmr": 16550,
-    "Ir": 1198932299,
-    "Overall": 1811358567
+    "D1mr": 4581665,
+    "D1mw": 2834520,
+    "DLmr": 1550571,
+    "DLmw": 1791810,
+    "Dr": 245534019,
+    "Dw": 95377424,
+    "I1mr": 4277300,
+    "ILmr": 15866,
+    "Ir": 932200278,
+    "Overall": 1420633071
 }
