+++ benchmarks/results/multithreading-1.json
@@ -1,12 +1,12 @@
 {
-    "D1mr": 2817562,
-    "D1mw": 540693,
-    "DLmr": 123338,
-    "DLmw": 121188,
-    "Dr": 108572041,
-    "Dw": 47853691,
-    "I1mr": 2467521,
-    "ILmr": 10249,
-    "Ir": 374192904,
-    "Overall": 561564990
+    "D1mr": 2808825,
+    "D1mw": 544188,
+    "DLmr": 122662,
+    "DLmw": 120659,
+    "Dr": 103979511,
+    "Dw": 43469910,
+    "I1mr": 2395315,
+    "ILmr": 9899,
+    "Ir": 357299406,
+    "Overall": 535338739
 }
