+++ benchmarks/results/pystone.json
@@ -1,12 +1,12 @@
 {
-    "D1mr": 6263814,
-    "D1mw": 1847167,
-    "DLmr": 92753,
-    "DLmw": 98979,
-    "Dr": 407397497,
-    "Dw": 181197841,
-    "I1mr": 12794252,
-    "ILmr": 10059,
-    "Ir": 1309531163,
-    "Overall": 1987801163
+    "D1mr": 5614673,
+    "D1mw": 1628287,
+    "DLmr": 92348,
+    "DLmw": 98625,
+    "Dr": 381027076,
+    "Dw": 155994843,
+    "I1mr": 9519850,
+    "ILmr": 9729,
+    "Ir": 1228557263,
+    "Overall": 1838651482
 }
