+++ filpreload/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "filpreload"
+version = "0.1.0"
+authors = ["Itamar Turner-Trauring <itamar@itamarst.org>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+lazy_static = "1.4.0"
+pymemprofile_api = {path = "../memapi"}
+parking_lot = "0.11.0"
+
+[target.'cfg(target_os = "linux")'.dependencies.jemallocator]
+version = "0.3.2"
+
+[lib]
+name = "filpreload"
+crate_type = ["staticlib"]
