+++ memapi/Cargo.toml
@@ -10,13 +10,9 @@ license = "Apache 2.0"
 [dependencies]
 im = "14.2.0"
 itertools = "0.8.2"
-lazy_static = "1.4.0"
 ahash = "0.6.3"
-parking_lot = "0.11.0"
 derivative = "2.1.1"
-
-[target.'cfg(target_os = "linux")'.dependencies.jemallocator]
-version = "0.3.2"
+lazy_static = "1.4.0"
 
 [target.'cfg(target_os = "linux")'.dependencies.cgroups-rs]
 version = "0.2.0"
@@ -35,7 +31,6 @@ proptest = "0.9.5"
 
 [lib]
 name = "pymemprofile_api"
-crate_type = ["staticlib", "rlib"]
 
 [features]
 default = []
